Create Statistical Test
During problem set 3, the student is prompted to plot the data and check if the data is normal, which I found interesting because in the lesson the instructor tells us that we should use a Shapiro-Wilk test to check if the data is normal. 

In the problem set, we plot the data and it looks quite skewed. On a succeeding quiz, we are asked to do a Mann-Whitney test to see if the samples(entries on rainy days vs. non-rainy days) is drawn from the same population. We get a 2% chance that these samples are drawn from the same population. Which means it is safe to say that significantly more people take the subway when it is raining.